pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-pytest"
version = "0.1.0"
description = "LLM-orchestrated testing framework for pytest"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Arno" }
]
keywords = ["testing", "llm", "pytest", "claude", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pytest>=7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40",
]
mcp = [
    "mcp>=1.0",
]
all = [
    "llm-pytest[browser,mcp]",
]

[project.entry-points.pytest11]
llm_pytest = "llm_pytest.plugin"

[project.scripts]
llm-pytest-mcp = "llm_pytest.mcp.base_server:main"

[tool.hatch.build.targets.wheel]
packages = ["llm_pytest"]
